refactor(ui): extract ChatInput component
